+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- A `get_submission()` method to download the submission of a single student.
+
+### Changed
+
+- The `get_submissions` method will now download submissions for _all_ students.
+
 ## [0.14.2] - 2026-01-22
 
 ### Fixed